+++ l10n_br_fiscal/models/document_abstract.py
@@ -6,38 +6,24 @@ from odoo import _, api, fields, models
 from odoo.exceptions import UserError
 from odoo.tools import DEFAULT_SERVER_DATETIME_FORMAT
 
-from ..constants.fiscal import (SITUACAO_EDOC,
-                                SITUACAO_EDOC_A_ENVIAR,
-                                SITUACAO_EDOC_AUTORIZADA,
-                                SITUACAO_EDOC_CANCELADA,
-                                SITUACAO_EDOC_DENEGADA,
-                                SITUACAO_EDOC_EM_DIGITACAO,
-                                SITUACAO_EDOC_ENVIADA,
-                                SITUACAO_EDOC_INUTILIZADA,
-                                SITUACAO_EDOC_REJEITADA, SITUACAO_FISCAL,
-                                SITUACAO_FISCAL_SPED_CONSIDERA_CANCELADO,
-                                TAX_FRAMEWORK,
-                                WORKFLOW_DOCUMENTO_NAO_ELETRONICO,
-                                WORKFLOW_EDOC,
+from ..constants.fiscal import (TAX_FRAMEWORK,
                                 PROCESSADOR)
-
-
-def fiter_nao_eletronico(record):
-    if not record.document_electronic:
-        return True
-    return False
 
 
 class DocumentAbstract(models.AbstractModel):
     _name = "l10n_br_fiscal.document.abstract"
-    _inherit = ["mail.thread", "mail.activity.mixin", "l10n_br_fiscal.document.mixin"]
+    _inherit = ["mail.thread",
+                "mail.activity.mixin",
+                "l10n_br_fiscal.document.mixin",
+                "l10n_br_fiscal.document.workflow"]
+
     _description = "Fiscal Document Abstract"
 
     """ Implementação base dos documentos fiscais
 
     Devemos sempre ter em mente que o modelo que vai usar este módulo abstrato
-     tem diversos metodos importantes e a intenção que os módulos da OCA que 
-     extendem este modelo, funcionem se possível sem a necessidade de 
+     tem diversos metodos importantes e a intenção que os módulos da OCA que
+     extendem este modelo, funcionem se possível sem a necessidade de
      codificação extra.
 
     É preciso também estar atento que o documento fiscal tem dois estados:
@@ -71,137 +57,6 @@ class DocumentAbstract(models.AbstractModel):
         self.amount_freight_value = sum(line.freight_value for line in self.line_ids)
         self.amount_total = sum(line.amount_total for line in self.line_ids)
 
-    @api.model
-    def _avaliable_transition(self, old_state, new_state):
-        """ Verifica as transições disponiveis, para não permitir alterações
-         de estado desconhecidas. Para mais detalhes verificar a variável
-          WORKFLOW_EDOC
-
-           (old_state, new_state) in (SITUACAO_EDOC_EM_DIGITACAO,
-                                      SITUACAO_EDOC_A_ENVIAR)
-
-        :param old_state: estado antigo
-        :param new_state: novo estado
-        :return:
-        """
-        if self.document_electronic:
-            return (old_state, new_state) in WORKFLOW_EDOC
-        else:
-            return (old_state, new_state) in WORKFLOW_DOCUMENTO_NAO_ELETRONICO
-
-    def _before_change_state(self, old_state, new_state):
-        """ Hook para realizar alterações antes da alteração do estado do edoc.
-
-        A variável self.state_edoc já estará com o estado antigo neste momento.
-
-        :param old_state:
-        :param new_state:
-        :return:
-        """
-        self.ensure_one()
-
-    def _exec_after_SITUACAO_EDOC_EM_DIGITACAO(self, old_state, new_state):
-        if self.state_fiscal in SITUACAO_FISCAL_SPED_CONSIDERA_CANCELADO:
-            raise (
-                _(
-                    "Não é possível retornar o documento para em \n"
-                    "digitação, quando o mesmo esta na situação: \n"
-                    "{0}, {1]".format(old_state, self.state_fiscal)
-                )
-            )
-
-    def _exec_after_SITUACAO_EDOC_A_ENVIAR(self, old_state, new_state):
-        pass
-
-    def _exec_after_SITUACAO_EDOC_ENVIADA(self, old_state, new_state):
-        pass
-
-    def _exec_after_SITUACAO_EDOC_REJEITADA(self, old_state, new_state):
-        pass
-
-    def _exec_after_SITUACAO_EDOC_AUTORIZADA(self, old_state, new_state):
-        pass
-
-    def _exec_after_SITUACAO_EDOC_CANCELADA(self, old_state, new_state):
-        pass
-
-    def _exec_after_SITUACAO_EDOC_DENEGADA(self, old_state, new_state):
-        pass
-
-    def _exec_after_SITUACAO_EDOC_INUTILIZADA(self, old_state, new_state):
-        pass
-
-    def _after_change_state(self, old_state, new_state):
-        """ Hook para realizar alterações depois da alteração do estado do doc.
-
-        A variável self.state_edoc já estará com o novo estado neste momento.
-
-        :param old_state:
-        :param new_state:
-        :return:
-        """
-        self.ensure_one()
-        if new_state == SITUACAO_EDOC_EM_DIGITACAO:
-            self._exec_after_SITUACAO_EDOC_EM_DIGITACAO(old_state, new_state)
-        elif new_state == SITUACAO_EDOC_A_ENVIAR:
-            self._exec_after_SITUACAO_EDOC_A_ENVIAR(old_state, new_state)
-        elif new_state == SITUACAO_EDOC_ENVIADA:
-            self._exec_after_SITUACAO_EDOC_ENVIADA(old_state, new_state)
-        elif new_state == SITUACAO_EDOC_REJEITADA:
-            self._exec_after_SITUACAO_EDOC_REJEITADA(old_state, new_state)
-        elif new_state == SITUACAO_EDOC_AUTORIZADA:
-            self._exec_after_SITUACAO_EDOC_AUTORIZADA(old_state, new_state)
-        elif new_state == SITUACAO_EDOC_CANCELADA:
-            self._exec_after_SITUACAO_EDOC_CANCELADA(old_state, new_state)
-        elif new_state == SITUACAO_EDOC_DENEGADA:
-            self._exec_after_SITUACAO_EDOC_DENEGADA(old_state, new_state)
-        elif new_state == SITUACAO_EDOC_INUTILIZADA:
-            self._exec_after_SITUACAO_EDOC_INUTILIZADA(old_state, new_state)
-
-    @api.multi
-    def _change_state(self, new_state):
-        """ Método para alterar o estado do documento fiscal, mantendo a
-        integridade do workflow da invoice.
-
-        Tenha muito cuidado ao alterar o workflow da invoice manualmente,
-        prefira alterar o estado do documento fiscal e ele se encarregar de
-        alterar o estado da invoice.
-
-        :param new_state: Novo estado
-        :return:
-        """
-
-        for record in self:
-            old_state = record.state_edoc
-
-            if not record._avaliable_transition(old_state, new_state):
-                raise UserError(
-                    _(
-                        "Não é possível realizar esta operação,\n"
-                        "esta transição não é permitida:\n\n"
-                        "De: {old_state}\n\n Para: {new_state}".format(
-                            old_state=old_state, new_state=new_state
-                        )
-                    )
-                )
-
-            record._before_change_state(old_state, new_state)
-            record.state_edoc = new_state
-            record._after_change_state(old_state, new_state)
-
-    state_edoc = fields.Selection(
-        selection=SITUACAO_EDOC,
-        string="Situação e-doc",
-        default=SITUACAO_EDOC_EM_DIGITACAO,
-        copy=False,
-        required=True,
-        track_visibility="onchange",
-        index=True,
-    )
-    state_fiscal = fields.Selection(
-        selection=SITUACAO_FISCAL, string="Situação Fiscal", track_visibility="onchange"
-    )
-
     processador_edoc = fields.Selection(
         string="Processador",
         selection=PROCESSADOR,
@@ -214,20 +69,6 @@ class DocumentAbstract(models.AbstractModel):
     active = fields.Boolean(
         string="Active",
         default=True)
-
-    number = fields.Char(
-        string="Number",
-        index=True)
-
-    key = fields.Char(
-        string="key",
-        index=True)
-
-    issuer = fields.Selection(
-        selection=[("company", "Company"), ("partner", "Partner")],
-        default="company",
-        required=True,
-        string="Issuer")
 
     user_id = fields.Many2one(
         comodel_name='res.users',
@@ -244,11 +85,7 @@ class DocumentAbstract(models.AbstractModel):
         related="document_type_id.electronic",
         string="Electronic?")
 
-    date = fields.Datetime(
-        string="Date")
-
-    date_in_out = fields.Datetime(
-        string="Date Move")
+    date_in_out = fields.Datetime(string="Date Move")
 
     document_serie_id = fields.Many2one(
         comodel_name="l10n_br_fiscal.document.serie",
@@ -477,75 +314,3 @@ class DocumentAbstract(models.AbstractModel):
                 ('active', '=', True)])
 
         self.document_serie_id = serie
-
-    def _date_server_format(self):
-        return fields.Datetime.now().strftime(DEFAULT_SERVER_DATETIME_FORMAT)
-
-    def document_date(self):
-        if not self.date:
-            self.date = self._date_server_format()
-
-    def document_number(self):
-        if not self.number and self.document_serie_id and self.date:
-            self.number = self._create_serie_number(
-                self.document_serie_id,
-                self.date
-            )
-
-    def document_check(self):
-        return True
-
-    def _document_confirm(self):
-        for record in self.filtered(fiter_nao_eletronico):
-            record._change_state(SITUACAO_EDOC_A_ENVIAR)
-
-    def action_document_confirm(self):
-        to_confirm = self.filtered(
-            lambda inv: inv.state_edoc != SITUACAO_EDOC_A_ENVIAR
-        )
-        # Este lambda acima evita erros quando o usuário clica várias vezes
-
-        to_confirm.document_date()
-        to_confirm.document_number()
-        to_confirm.document_check()
-
-        return to_confirm._document_confirm()
-
-    def _document_send(self):
-        for record in self.filtered(fiter_nao_eletronico):
-            self._change_state(SITUACAO_EDOC_AUTORIZADA)
-
-    def action_document_send(self):
-        to_confirm = self.filtered(
-            lambda inv: inv.state_edoc in (SITUACAO_EDOC_EM_DIGITACAO,
-                                           SITUACAO_EDOC_REJEITADA))
-        if to_confirm:
-            to_confirm.action_document_confirm()
-
-        to_send = self.filtered(
-            lambda inv: inv.state_edoc == SITUACAO_EDOC_A_ENVIAR
-        )
-        return to_send._document_send()
-
-    def _document_cancel(self, justificative):
-        self._change_state(SITUACAO_EDOC_CANCELADA)
-        msg = "Cancelamento: {}".format(justificative)
-        self.message_post(body=msg)
-
-    @api.multi
-    def action_document_cancel(self):
-        result = self.env["ir.actions.act_window"].for_xml_id(
-            "l10n_br_fiscal", "wizard_document_cancel_action"
-        )
-        return result
-
-    def _document_correction(self, justificative):
-        msg = "Carta de correção: {}".format(justificative)
-        self.message_post(body=msg)
-
-    @api.multi
-    def action_document_correction(self):
-        result = self.env["ir.actions.act_window"].for_xml_id(
-            "l10n_br_fiscal", "wizard_document_correction_action"
-        )
-        return result
